Parameters   -renamed field_face_rank -> min_face_rank   -added Method:gravity_mg:level_[min|max] parameter   -changed gravity_mg:iter_max to 10 Data   -renamed mesh_Data.?pp to data_Data.?pp Adapt   -removed unused FACE_LEVEL from PUT_LEVEL macro   -tweaking RefineSlope (noticed anamalous behavior in meshed-up.in    with block size 8x8--still an issue (bug #46)) Global   -addressed compiler warnings Mesh   -added missing initializations in Block::Block()   -bug fix: removed Block::refresh_index_ (redundant with index_refresh_) Problem   -added missing initializations in Problem::Problem() Gravity   -implemented EnzoComputeSmoothJacobi::compute_()  (untested) Testing   -removed NeighborLB tests from Balance since apparently not doing anything   -added meshed-up.[in|png]

+++ src/Cello/mesh_ItNeighbor.hpp
@@ -109,10 +109,10 @@ private: // attributes
   int ipf3_[3];
 
   /// simulation rank
-  size_t rank_;
+  int rank_;
 
   /// face rank limit
-  size_t min_face_rank_;
+  int min_face_rank_;
 
   /// Whether domain is periodic along each face & axis
   bool periodic_[3][2];
